+++ home/home.nix
@@ -4,7 +4,7 @@
     ./modules/git.nix
     ./modules/niri.nix
     ./modules/noctalia.nix
-  ]
+  ];
 
   home.username = "thedreamdev";
   home.homeDirectory = "/home/thedreamdev";