+++ CHANGELOG.md
@@ -1,3 +1,3 @@
 # Version 26.5.18.3
 
-- Installed Niri and Noctalia
+- Configuring Niri and Noctalia